FEATURE/product/S3 object key 생성 및 업로드 URL 발급 기능 추가
