+++ src/main/java/com/ezmeal/product/application/response/ProductImageUploadUrlResponse.java
@@ -1,0 +1,10 @@
+package com.ezmeal.product.application.response;
+
+public record ProductImageUploadUrlResponse(
+        String uploadUrl,
+        String objectKey,
+        String method,
+        String contentType,
+        Long expiresInSeconds
+) {
+}
