+++ src/main/java/com/ezmeal/product/application/upload/PresignedUrlProvider.java
@@ -1,0 +1,5 @@
+package com.ezmeal.product.application.upload;
+
+public interface PresignedUrlProvider {
+    String createUploadUrl(String objectKey, String contentType, long expiresInSeconds);
+}
