+++ src/main/java/com/ezmeal/product/application/upload/ProductImageUploadType.java
@@ -1,0 +1,6 @@
+package com.ezmeal.product.application.upload;
+
+public enum ProductImageUploadType {
+    PRODUCT_MAIN_IMAGE,
+    MEAL_PLAN_IMAGE
+}
