+++ src/main/java/com/ezmeal/product/infrastructure/search/elasticsearch/indexer/ProductSearchDocumentRepository.java
@@ -1,0 +1,9 @@
+package com.ezmeal.product.infrastructure.search.elasticsearch.indexer;
+
+import com.ezmeal.product.infrastructure.search.elasticsearch.document.ProductSearchDocument;
+import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
+
+public interface ProductSearchDocumentRepository extends ElasticsearchRepository<ProductSearchDocument, String> {
+
+
+}
