+++ src/main/java/com/ezmeal/product/infrastructure/search/elasticsearch/log/ElasticSearchProductSearchLogAppender.java
@@ -1,0 +1,20 @@
+package com.ezmeal.product.infrastructure.search.elasticsearch.log;
+
+import com.ezmeal.product.application.search.log.ProductSearchLogAppender;
+import com.ezmeal.product.application.search.log.ProductSearchLogCommand;
+import com.ezmeal.product.infrastructure.search.elasticsearch.document.ProductSearchLogDocument;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class ElasticSearchProductSearchLogAppender implements ProductSearchLogAppender {
+
+    private final ProductSearchLogDocumentRepository productSearchLogDocumentRepository;
+
+    @Override
+    public void append(ProductSearchLogCommand command) {
+        ProductSearchLogDocument document = ProductSearchLogDocument.from(command);
+        productSearchLogDocumentRepository.save(document);
+    }
+}
