--- src/main/java/com/ezmeal/product/infrastruture/persistence/companySnapshotRepository/JpaCompanySnapshotRepository.java
@@ -1,18 +1,0 @@
-package com.ezmeal.product.infrastruture.persistence.companySnapshotRepository;
-
-import com.ezmeal.product.domain.model.companySnapShot.CompanySnapshot;
-import java.util.Optional;
-import java.util.UUID;
-import org.springframework.data.jpa.repository.EntityGraph;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface JpaCompanySnapshotRepository extends JpaRepository<CompanySnapshot, UUID> {
-
-    Optional<CompanySnapshot> findByCompanyId(UUID companyId);
-
-    Optional<CompanySnapshot> findByCompanyIdAndDeletedAtIsNull(UUID companyId);
-
-    @EntityGraph(attributePaths = "deliveryAreas")
-    Optional<CompanySnapshot> findWithDeliveryAreasByCompanyIdAndDeletedAtIsNull(UUID companyId);
-
-}
