--- src/main/java/com/ezmeal/product/infrastruture/search/elasticsearch/indexer/ProductSearchDocumentRepository.java
@@ -1,9 +1,0 @@
-package com.ezmeal.product.infrastruture.search.elasticsearch.indexer;
-
-import com.ezmeal.product.infrastruture.search.elasticsearch.document.ProductSearchDocument;
-import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
-
-public interface ProductSearchDocumentRepository extends ElasticsearchRepository<ProductSearchDocument, String> {
-
-
-}
