--- src/main/java/com/ezmeal/product/infrastruture/search/elasticsearch/log/ProductSearchLogDocumentRepository.java
@@ -1,8 +1,0 @@
-package com.ezmeal.product.infrastruture.search.elasticsearch.log;
-
-import com.ezmeal.product.infrastruture.search.elasticsearch.document.ProductSearchLogDocument;
-import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
-
-public interface ProductSearchLogDocumentRepository extends ElasticsearchRepository<ProductSearchLogDocument, String> {
-
-}
